Install packages as available on set date - groundhog.library()

Install packages as available on set date - groundhog.library()

groundhog.library(
  pkg,
  date,
  quiet.install = TRUE,
  include.suggests = FALSE,
  ignore.deps = c(),
  force.source = FALSE,
  force.install = FALSE
)

Arguments

pkg
character string pr vector with name of target package(s) to load (and install if needed), need not be in quotes.

date
character string (yyyy-mm-dd), or date value, with the date which determines the version of the package, and all dependencies, to be loaded (and installed if needed).

quiet.install
logical, defaults to TRUE. When set to FALSE, displays output generated by install.packages() when installing from source

include.suggests
logical, defaults to FALSE. When set to TRUE, installs/ loads dependencies classified in the DESCRIPTION file as suggested.

ignore.deps
an optional character vector containing dependencies which may be already loaded in the R session and even if the loaded version does not match the version implied by the entered date, groundhog.library() will proceed and ignore this conflict. If one version of a package is loaded, and a different is needed for groundhog, the default behavior is to stop the request and ask the user to restart the R session to unload all packages. This will bypass that requirement.

force.source
logical (defaults to FALSE). When set to TRUE, will not attempt installing binary from CRAN or MRAN and instead download source file and install it.

force.install
logical (defaults to FALSE). When set to TRUE, will delete existing package files in groundhog folder, and install anew.

Value

a character vector containing all active packages for the session, with their version number, under the format pkg_vrs.

Examples

if (FALSE) {
groundhog.library("magrittr", "2020-07-12")

pkgs <- c('pwr','metafor')
groundhog.library(pkgs, "2020-02-12")
}