Install package from source

Install package from source

install.source(
  pkg_vrs,
  lib,
  date,
  force.download = FALSE,
  quiet.install = FALSE
)

Arguments

pkg_vrs
character. The required package and its version separated by an underscore.

lib
character. Path where to install the package.

date
date (in the format "%Y-%m-%d"), Required date for the package.

force.download
logical (defaults to FALSE). If the tarball containing the package source code already exists locally, should it be downloaded again?

quiet.install
Logical (defaults to TRUE).Run install.packages() with quiet = TRUE?

Value

Invisible NULL.